Amends in documentation and migration

+++ db/migrate/20170930000001_create_post_types.rb
@@ -1,28 +1,26 @@
 class CreatePostTypes < ActiveRecord::Migration[5.1]
   def up
-    unless PostType.table_exists?
-      create_table :post_types do |t|
-        t.timestamps
-        t.boolean :active, default: true, null: false
-        t.integer :posts_count, default: 0, null: false
-        t.integer :category_depth, limit: 2, default: 0
-        t.string :name, null: false
-        t.string :slug, null: false
-        t.string :default_category_name
-      end
+    return if PostType.table_exists?
 
-      add_index :post_types, :slug, unique: true
-      add_index :post_types, :name, unique: true
-
-      create_default_types
-      create_privileges
+    create_table :post_types do |t|
+      t.timestamps
+      t.boolean :active, default: true, null: false
+      t.integer :posts_count, default: 0, null: false
+      t.integer :category_depth, limit: 2, default: 0
+      t.string :name, null: false
+      t.string :slug, null: false
+      t.string :default_category_name
     end
+
+    add_index :post_types, :slug, unique: true
+    add_index :post_types, :name, unique: true
+
+    create_default_types
+    create_privileges
   end
 
   def down
-    if PostType.table_exists?
-      drop_table :post_types
-    end
+    drop_table :post_types if PostType.table_exists?
   end
 
   def create_default_types
@@ -32,7 +30,7 @@ class CreatePostTypes < ActiveRecord::Migration[5.1]
   end
 
   def create_privileges
-    group        = PrivilegeGroup.find_by!(slug: 'editors')
+    group        = PrivilegeGroup.find_by(slug: 'editors') || PrivilegeGroup.create!(slug: 'editors', name: 'Редакторы')
     chief_editor = Privilege.find_by(slug: 'chief_editor') || Privilege.create(slug: 'chief_editor', name: 'Главный редактор')
     children     = {
       editor:   'Редактор',
@@ -43,9 +41,7 @@ class CreatePostTypes < ActiveRecord::Migration[5.1]
     children.each do |slug, name|
       child = Privilege.new(parent: chief_editor, slug: slug, name: name, administrative: false)
 
-      if child.save
-        group.add_privilege(child)
-      end
+      group.add_privilege(child) if child.save
     end
   end
 end
